Rust: Highlight byte-string literals.
This adds support for two new string literals and one new string literal,
corresponding to the existing versions of these. Compared to the originals, the
new literals have slightly different escapes and are ASCII only. I've decided
to simply add flags to the existing scanners to handle them. New styles had to
be added to handle the line spanning string literals. The byte character style
was added for consistency.