Backport: Bug [#2095]. gtk: Accessible: emit the text deletion signal before deletion
This allows the accessibility layer to request detail, such as which
characters have been deleted.  ATK will query the deleted range to
report the data that was removed, so it must still be available for
the query to give a correct answer.
So, emit the signal in BEFOREDELETE instead of DELETETEXT.

Backport of changeset 7406:0a937b4c0b2b.